Split the Zarr extension into a directory
Mirrors ext/NumericalEarthArchGDALExt: the module file carries the imports and
includes, dataset code lives in its own file, and the shared filter and window
helpers sit alongside it.
